rename file "ecco.h" to "ECCO.h"
