+++ pkg/ctrl/ctrl_cost_gen.F
@@ -51,7 +51,7 @@ c     == global variables ==
 
 #ifdef ALLOW_ECCO
 #  include "ECCO_SIZE.h"
-#  include "ecco.h"
+#  include "ECCO.h"
 #endif
 #ifdef ALLOW_CTRL
 # include "ctrl.h"
@@ -439,7 +439,7 @@ c     == global variables ==
 
 #ifdef ALLOW_ECCO
 #  include "ECCO_SIZE.h"
-#  include "ecco.h"
+#  include "ECCO.h"
 #endif
 #ifdef ALLOW_CTRL
 # include "ctrl.h"
