+++ pkg/ecco/cost_averagesfields.F
@@ -45,7 +45,7 @@ c     == global variables ==
 
 #ifndef ECCO_CTRL_DEPRECATED
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #else
 # include "ecco_cost.h"
 # ifdef ALLOW_CTRL
@@ -294,7 +294,7 @@ cph#ifdef ALLOW_SEAICE_COST_AREASST
      &     sstbar, theta(1-Olx,1-Oly,1,1,1), xx_sstbar_mean_dummy,
      &     first, last, startofday, endofday, inday,
      &     sum1day, dayrec, 1, mythid )
-      ENDIF !      IF (using_cost_sst) THEN  
+      ENDIF !      IF (using_cost_sst) THEN
 #endif
 
 #if (defined (ALLOW_SALT_COST_CONTRIBUTION) || \
@@ -453,7 +453,6 @@ c--     Save ubar and vbar.
 
         call active_write_xy( fnametauybar, tauybar, monrec, optimcycle,
      &                         mythid, xx_tauy_mean_dummy)
-
 
       else if ( .not. (first.or. startofmonth) .and.
      &          .not. (last .or. endofmonth)       ) then
