+++ pkg/ecco/cost_bp_read.F
@@ -31,7 +31,7 @@ c     == global variables ==
 #endif
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 
 c     == routine arguments ==
