+++ pkg/ecco/cost_gencost_all.F
@@ -22,7 +22,7 @@ c     == global variables ==
 #include "GRID.h"
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 
 c     == routine arguments ==
