+++ pkg/ecco/cost_gencost_assignperiod.F
@@ -21,7 +21,7 @@ c     == global variables ==
 #include "PARAMS.h"
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 
 c     == routine arguments ==
