+++ pkg/ecco/cost_gencost_bpv4.F
@@ -29,7 +29,7 @@ c     == global variables ==
 
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 
 c     == routine arguments ==
