+++ pkg/ecco/cost_gencost_customize.F
@@ -30,7 +30,7 @@ c     == global variables ==
 #include "FFIELDS.h"
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 #ifdef ALLOW_SEAICE
 #  include "SEAICE_SIZE.h"
