+++ pkg/ecco/cost_gencost_moc.F
@@ -27,7 +27,7 @@ c     == global variables ==
 #include "GRID.h"
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 
 c     == routine arguments ==
