+++ pkg/ecco/cost_gencost_sshv4.F
@@ -41,7 +41,7 @@ c     == global variables ==
 #endif
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 #ifdef ALLOW_SMOOTH
 # include "SMOOTH.h"
