+++ pkg/ecco/cost_gencost_sstv4.F
@@ -30,7 +30,7 @@ c     == global variables ==
 #endif
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 #ifdef ALLOW_SMOOTH
 # include "SMOOTH.h"
