+++ pkg/ecco/cost_gencost_transp.F
@@ -25,7 +25,7 @@ c     == global variables ==
 #endif
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 
 c     == routine arguments ==
@@ -73,7 +73,6 @@ c- facNum is 1 divided by the number of tiles in SIZE dot h
       _RL preproc_r(NGENPPROC)
       character*(MAX_LEN_FNAM) preproc(NGENPPROC)
       character*(MAX_LEN_FNAM) preproc_c(NGENPPROC)
-
 
       logical doglobalread
       logical ladinit
@@ -140,7 +139,7 @@ c--   Initialise local variables.
         nrecloc=0
         nrecloc=gencost_nrec(kgen(icount))
 
-c-- only enters if there is at least 1 record 
+c-- only enters if there is at least 1 record
         if(nrecloc.gt.0) then
 
           facCost=1. _d 0 / float(nrecloc)
@@ -167,7 +166,7 @@ c to work for comparison of characters so match dosumsq for now.
 
 c-- report of dosumsq flag to make sure it is false
           il=ilnblnk(gencost_name(kgen(icount)))
-          write(msgbuf,'(3A,L5,2A)') 
+          write(msgbuf,'(3A,L5,2A)')
      &    'Cost ',gencost_name(kgen(icount))(1:il),
      &    ' dosumsq: ',dosumsq,' preproc(1): ',preproc(1)(1:7)
           CALL PRINT_MESSAGE( msgBuf, standardMessageUnit,
@@ -180,11 +179,11 @@ c-- set obsfile if defined. Not use for now. send warning
           doreadobs=.FALSE.
           if( .not. gencost_datafile(kgen(icount)).eq.' ') then
 c            doreadobs=.TRUE.
-            write(msgbuf,'(A)') 
+            write(msgbuf,'(A)')
      &      '**WARNING** S/R COST_GENCOST_TRANSP: gencost_datafile '
             CALL PRINT_MESSAGE( msgBuf, errorMessageUnit,
      &                          SQUEEZE_RIGHT, myThid )
-            write(msgbuf,'(A)') 
+            write(msgbuf,'(A)')
      &      'are currently ignored. Adjust the S/R to add code.'
             CALL PRINT_MESSAGE( msgBuf, errorMessageUnit,
      &                          SQUEEZE_RIGHT, myThid )
@@ -215,7 +214,6 @@ c gencost_mask and fname0 are dummy, get fname1 from mybarfile
      &       nrecloc,preproc,preproc_c,preproc_i,preproc_r,
      &       dummy,mythid)
 
-
 c--------- PART 1.2 read data --------------------
 
 c-- ignore for now, but use doreadobs flag if needed
@@ -224,9 +222,7 @@ c            if(doreadobs) then
             call ecco_zero(myobs,Nr,zeroRL,myThid)
 c            endif
 
-
 c--------- PART 1.3 Cost calculation -------------
-
 
 c-- keep total at each irec to print out for time-series
             DO bj = jtlo,jthi
@@ -236,7 +232,7 @@ c-- keep total at each irec to print out for time-series
               ENDDO
             ENDDO
 
-c compute obs minus bar (localdif) and mask (difmask) 
+c compute obs minus bar (localdif) and mask (difmask)
 c note localtmp is set to 1.
             call ecco_zero(localtmp,nnzobs,oneRL,mythid)
             call ecco_zero(localdif,nnzobs,zeroRL,mythid)
@@ -287,7 +283,7 @@ c sum that is actually be used in cost function
 
           enddo !irec
 
-c-- last step: 
+c-- last step:
 c-- divide by number of record to get mean transport:
 c-- make num_gencost equals number of months/days used
           do bj = jtlo,jthi
