+++ pkg/ecco/cost_generic.F
@@ -35,7 +35,7 @@ c     == global variables ==
 #endif
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 #ifdef ALLOW_SEAICE
 # include "SEAICE_COST.h"
@@ -185,7 +185,7 @@ c     == global variables ==
 #endif
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 #ifdef ALLOW_SEAICE
 # include "SEAICE_COST.h"
