+++ pkg/ecco/cost_genread.F
@@ -31,7 +31,7 @@ C     == global variables ==
 #include "SIZE.h"
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 
 c     == routine arguments ==
