+++ pkg/ecco/cost_sla_read.F
@@ -28,7 +28,7 @@ c     == global variables ==
 
 #include "cal.h"
 #include "ECCO_SIZE.h"
-#include "ecco.h"
+#include "ECCO.h"
 
 c     == routine arguments ==
