+++ pkg/ecco/ecco_check.F
@@ -741,7 +741,7 @@ C     === Global variables ===
 #include "EEPARAMS.h"
 #include "PARAMS.h"
 #include "ECCO_SIZE.h"
-#include "ecco.h"
+#include "ECCO.h"
 #ifdef ALLOW_CAL
 # include "cal.h"
 #endif
