+++ pkg/ecco/ecco_cost_driver.F
@@ -27,7 +27,7 @@ c     == global variables ==
 #include "PARAMS.h"
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 #ifdef ALLOW_AUTODIFF_TAMC
 # ifdef ALLOW_SSH_COST_CONTRIBUTION
