+++ pkg/ecco/ecco_cost_init_barfiles.F
@@ -35,7 +35,7 @@ c     == global variables ==
 #  include "ecco_cost.h"
 # else
 #  include "ECCO_SIZE.h"
-#  include "ecco.h"
+#  include "ECCO.h"
 # endif
 #endif
 #ifdef ALLOW_CTRL
