+++ pkg/ecco/ecco_phys.F
@@ -25,7 +25,7 @@ c     == global variables ==
 #include "GRID.h"
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 #ifdef ALLOW_PTRACERS
 # include "PTRACERS_SIZE.h"
