+++ pkg/ecco/ecco_summary.F
@@ -32,7 +32,7 @@ c     == global variables ==
 #include "EEPARAMS.h"
 #include "SIZE.h"
 #include "ECCO_SIZE.h"
-#include "ecco.h"
+#include "ECCO.h"
 
 #ifdef ALLOW_CTRL
 # include "optim.h"
