+++ pkg/ecco/stergloh_output.F
@@ -21,7 +21,7 @@ C     === Global variables ===
 #include "PARAMS.h"
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 
 C !INPUT PARAMETERS:
