+++ pkg/seaice/seaice_cost_sss.F
@@ -38,7 +38,7 @@ c     == global variables ==
 # endif
 # ifdef ALLOW_ECCO
 #  include "ECCO_SIZE.h"
-#  include "ecco.h"
+#  include "ECCO.h"
 # endif
 # ifdef ALLOW_SEAICE
 #  include "SEAICE_COST.h"
