+++ pkg/seaice/seaice_cost_weights.F
@@ -6,7 +6,7 @@
 #include "CTRL_OPTIONS.h"
 #endif
 
-      subroutine seaice_cost_weights( mythid )
+      subroutine seaice_cost_weights( myThid )
 
 c     ==================================================================
 c     SUBROUTINE seaice_cost_weights
@@ -30,13 +30,13 @@ c     == global variables ==
 #endif
 #ifdef ALLOW_ECCO
 # include "ECCO_SIZE.h"
-# include "ecco.h"
+# include "ECCO.h"
 #endif
 #include "SEAICE_COST.h"
 
 c     == routine arguments ==
 
-      integer  mythid
+      integer  myThid
 
 #ifdef ALLOW_ECCO
 c     == local variables ==
@@ -62,21 +62,21 @@ c     == external ==
 
 c     == end of interface ==
 
-      jtlo = mybylo(mythid)
-      jthi = mybyhi(mythid)
-      itlo = mybxlo(mythid)
-      ithi = mybxhi(mythid)
-      jmin = 1-oly
-      jmax = sny+oly
-      imin = 1-olx
-      imax = snx+olx
+      jtlo = myByLo(myThid)
+      jthi = myByHi(myThid)
+      itlo = myBxLo(myThid)
+      ithi = myBxHi(myThid)
+      jmin = 1-OLy
+      jmax = sNy+OLy
+      imin = 1-OLx
+      imax = sNx+OLx
 
 c--   Define frame.
       do j = jmin,jmax
         do i = imin,imax
 c--       North/South and West/East edges set to zero.
-          if ( (j .lt. 1) .or. (j .gt. sny) .or.
-     &         (i .lt. 1) .or. (i .gt. snx)      ) then
+          if ( (j .lt. 1) .or. (j .gt. sNy) .or.
+     &         (i .lt. 1) .or. (i .gt. sNx)      ) then
             frame(i,j) = 0. _d 0
           else
             frame(i,j) = 1. _d 0
@@ -101,7 +101,7 @@ c--
       k     = 1
       if ( smrarea_errfile .NE. ' ' ) then
          CALL READ_REC_3D_RL( smrarea_errfile, cost_iprec, nnz,
-     &                        wsmrarea, irec, 0, mythid )
+     &                        wsmrarea, irec, 0, myThid )
          do bj = jtlo,jthi
            do bi = itlo,ithi
              do j = jmin,jmax
@@ -138,7 +138,7 @@ c--
       enddo
 #endif /* ALLOW_SEAICE_COST_SMR_AREA */
 
-#endif
+#endif /* ALLOW_ECCO */
 
-      return
-      end
+      RETURN
+      END
